+++ R/na_summary.R
@@ -2,7 +2,6 @@
 #' @inheritParams percent_missing
 #' @inheritParams sort_by_missingness
 #' @importFrom stats "aggregate" "as.formula" "na.pass"
-#' @param ... Arguments to other functions
 #' @examples
 #' na_summary(airquality)
 #' # grouping
@@ -16,7 +15,7 @@
 #'
 #' @export
 
-na_summary <- function(df,grouping_cols=NULL,sort_by=NULL,descending=FALSE,...){
+na_summary <- function(df,grouping_cols=NULL,sort_by=NULL,descending=FALSE, exclude_cols = NULL){
   UseMethod("na_summary")
 
 
@@ -24,30 +23,26 @@ na_summary <- function(df,grouping_cols=NULL,sort_by=NULL,descending=FALSE,...){
 
 #' @export
 
-na_summary.data.frame <- function(df,grouping_cols=NULL,sort_by=NULL,descending=FALSE,...){
+na_summary.data.frame <- function(df,grouping_cols=NULL,sort_by=NULL,descending=FALSE,
+                                  exclude_cols= NULL){
   # stick to(with?) base as much as possible
   # get total NAs columnwise
-all_counts <-stack(get_na_counts(df))
-all_percents <- stack(percent_missing(df,...))
+all_counts <-stack(get_na_counts(df, grouping_cols = grouping_cols, exclude_cols = exclude_cols))
+all_percents <- stack(percent_missing(df,grouping_cols = grouping_cols, exclude_cols = exclude_cols))
 names(all_counts) <- c("missing","variable")
 names(all_percents) <- c("percent_missing","variable")
 if(is.null(grouping_cols)){
 
-if(nrow(all_counts) != nrow(all_percents)){
+  if(nrow(all_counts) != nrow(all_percents)){
   stop("Binding of datasets failed. Please check using percent_missing and get_na_counts first")
 }
 
 
+all_counts$complete <- ifelse(all_counts$missing==0,nrow(df),nrow(df) - all_counts$missing)
 
+all_counts$percent_complete <- ifelse(all_percents$percent_missing==0,100,100 - all_percents$percent_missing)
 
-  all_counts$complete <- ifelse(all_counts$missing==0,nrow(df),
-                                nrow(df) - all_counts$missing)
-
-  all_counts$percent_complete <- ifelse(all_percents$percent_missing==0,
-                                       100,
-                                        100 - all_percents$percent_missing)
-
-  res <- merge(all_counts,all_percents,by="variable")
+res <- merge(all_counts,all_percents,by="variable")
 
 }
 
@@ -56,23 +51,17 @@ if(nrow(all_counts) != nrow(all_percents)){
 else{
     non_grouping = setdiff(names(df), grouping_cols)
     #matched_groups = which(names(df) %in% grouping_cols)
-    if(length(non_grouping) > 1){
-      warning("All non grouping values used. Using select non groups is currently not supported")
-    }
-    if(!all(grouping_cols %in% names(df))){
-      stop("All grouping_cols should exist in the dataset.")
-    }
+    if(length(non_grouping) > 1) warning("All non grouping values used. Using select non groups is currently not supported")
+
+    if(!all(grouping_cols %in% names(df))) stop("All grouping_cols should exist in the dataset.")
     grouping_cols_formula = paste0(grouping_cols,collapse="+")
     agg_formula <- as.formula(paste0(".~",
                                      grouping_cols_formula))
     res<-do.call(data.frame,aggregate(agg_formula,data=df,
-                                      function(x)
-               c(missing = sum(is.na(x)),
-        complete = length(x) - sum(is.na(x)),
-        percent_complete = sum(!is.na(x)) / length(x) * 100,
-        percent_missing = sum(is.na(x)) / length(x) * 100
-        ) ,
-      na.action = na.pass)) %>%
+                                      function(x) c(missing = sum(is.na(x)), complete = length(x) - sum(is.na(x)),
+                             percent_complete = sum(!is.na(x)) / length(x) * 100,
+                             percent_missing = sum(is.na(x)) / length(x) * 100
+        ) , na.action = na.pass)) %>%
     tidyr::pivot_longer(cols = -all_of(grouping_cols)) %>%
       tidyr::separate(name,c("variable","metric"),sep="\\.(?=percent|miss|complete)")  %>%
       tidyr::pivot_wider(names_from=metric,values_from=value)