+++ tests/testthat/test_na_summary.R
@@ -3,12 +3,10 @@ testthat::test_that(desc="Test na_summary",
 
                       skip_on_oldrel()
 
-     expect_warning(mde::na_summary(airquality, grouping_cols=c("Month",
-                                                                          "Day")),
+     expect_warning(na_summary(airquality, grouping_cols=c("Month","Day")),
                               "All non grouping values used. Using select non groups is currently not supported",
                               fixed=TRUE)
-    expect_error(mde::na_summary(airquality,grouping_cols="gibberish"),
-                           "All grouping_cols should exist in the dataset.",
+    expect_error(na_summary(airquality,grouping_cols="gibberish"), "All grouping_cols should exist in the dataset.",
                            fixed=TRUE)
     # expect that all rows are equal as required
     expect_error(na_summary(mtcars, exclude_cols = "mpg"),